feat: Add velocity property to CreatureBundle
This commit adds the `velocity` property to the `CreatureBundle` struct in the `creature.rs` file. The `velocity` property is of type `Velocity` and is used to store the velocity of the creature. This change allows for more accurate movement calculations and enhances the overall behavior of the creature in the game.